fix: 改用 FP32 精度解决 GPU 编译错误
- model_export_utils.py: 移除 .half() 和 dtype=torch.float16
- model_manager.py: 移除 compress_to_fp16=True
- subgraph_executor.py: 改用 astype(np.float32)
- executor.py, pipeline_executor.py: 输出 tensor 改用 float32
- openvino_async_wrapper.py: 输入转换改用 float32

匹配 profiling 代码格式，修复 GPU 编译时 F16→F32 转换缺少 kernel 的问题

+++ executer/executor.py
@@ -173,10 +173,10 @@ class PipelineExecutor:
         print(f"Execution Mode: {'Pipeline Parallel' if use_pipeline_parallelism else 'Sequential'}")
         print(f"{'='*70}\n")
 
-        # 准备输出（使用 FP16 以匹配模型输出精度）
+        # 准备输出（使用 FP32 以匹配 OpenVINO GPU 输出精度）
         total_nodes = self.data_loader.full_data.num_nodes
         hidden_dim = 256  # GraphSAGE默认hidden dimension
-        all_embeddings = torch.zeros(total_nodes, hidden_dim, dtype=torch.float16)
+        all_embeddings = torch.zeros(total_nodes, hidden_dim, dtype=torch.float32)
 
         per_subgraph_times = []
         per_cluster_times = []
